server update

+++ database/migrations/2020_09_27_161239_create_products_table.php
@@ -23,7 +23,7 @@ class CreateProductsTable extends Migration
             $table->double('product_unit', 12,4)->default(0)->comment('piece = 0, dozen = 1');
             $table->double('purchase_price', 12,4)->default(0);
             $table->double('sale_price', 12,4)->default(0);
-            $table->double('warning_quantity', 12,4)->default(0);
+            $table->integer('warning_quantity')->default(0);
             $table->smallInteger('tax_percentage')->default(0)->comment('percentage(%)');
             $table->string('product_detail')->nullable();
             $table->string('image', 30)->nullable();
